fix bugs

+++ Core/views.py
@@ -11,10 +11,11 @@ from django.utils.timezone import make_aware
 from django_filters.rest_framework import DjangoFilterBackend
 from django.db.models import F,Q,Subquery,OuterRef
 import string
-
-from .models import User,Task,TeamMembership,Team
-from .serializers import UserSerializer,TeamMembershipSerializer,TeamSerializer
+import os
+from .models import User,Task,TeamMembership,Team,Document
+from .serializers import UserSerializer,TeamMembershipSerializer,TeamSerializer,DocumentSerializer,TaskSerializer
 from .permissions import IsAdministrater,IsTeamAdministrator,IsTeamCreater,IsTeamMember
+from .support import move_files,move_file
 
 from rest_framework import generics,viewsets,permissions,status
 from rest_framework.generics import CreateAPIView
@@ -22,7 +23,7 @@ from rest_framework.authtoken.models import Token
 from rest_framework.views import APIView
 from rest_framework.decorators import action
 from rest_framework.response import Response
-from rest_framework.exceptions import NotAuthenticated,AuthenticationFailed
+from rest_framework.exceptions import NotAuthenticated,AuthenticationFailed,PermissionDenied
 from rest_framework.permissions import AllowAny,IsAuthenticated
 
 # Create your views here.
@@ -91,6 +92,12 @@ class TeamManagerView(viewsets.ModelViewSet):
             TeamMembership.objects.create(user=request.user,team=serializer.instance,role='Creater',permission='rw')
             return Response({"message":"success"}, status=200)
         return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
+    
+    def list(self,request):
+        user_teams = Team.objects.filter(users=request.user)
+        return Response(TeamSerializer(user_teams,many=True).data,status=200)
+        
+        
 
 
     
@@ -113,9 +120,10 @@ class InviteView(viewsets.ModelViewSet):
             if TeamMembership.objects.filter(user=user,team=team).exists():
                 return Response({"message":"user already exists"}, status=status.HTTP_400_BAD_REQUEST)
             else:
-                TeamMembership.objects.create(user=user,team=team,role='Viewer',permission='r')
-                return Response({"message":"success"}, status=200)
-        raise AuthenticationFailed()
+                temp = TeamMembership.objects.create(user=user,team=team,role='Viewer',permission='r')
+                print(TeamMembershipSerializer(temp).data)
+                return Response(TeamMembershipSerializer(temp).data, status=200)
+        raise PermissionDenied()
 
     
     #找所有team_id对应的成员
@@ -126,7 +134,7 @@ class InviteView(viewsets.ModelViewSet):
             teammembers = TeamMembership.objects.filter(team_id=team_id)
             return Response(TeamMembershipSerializer(teammembers, many=True).data, status=200)
             
-        raise AuthenticationFailed()
+        raise PermissionDenied()
     
 
     
@@ -134,15 +142,19 @@ class InviteView(viewsets.ModelViewSet):
     def delete(self,request):
         if(IsTeamAdministrator(request)):
             user_id = request.data.get('user_id')
+            username = request.data.get('username')
             team_id = request.data.get('team_id')
-            teamMembership = TeamMembership.objects.filter(user_id=user_id,team_id=team_id).first()
-            print(teamMembership)
+            user = User.objects.filter(Q(id=user_id) | Q(username=username)).first()
+            teamMembership = TeamMembership.objects.filter(user=user,team_id=team_id).first()
+            if not teamMembership:
+                return Response({"message":"user or team not found"}, status=status.HTTP_400_BAD_REQUEST)
+
             if teamMembership.role == "Viewer":
                 teamMembership.delete()
                 return Response({"message":"success"}, status=200)
             else:
                 return Response({"message":"cannot delete creater or administrator"}, status=status.HTTP_400_BAD_REQUEST)
-        raise AuthenticationFailed()
+        raise PermissionDenied()
         
         
     
@@ -160,12 +172,12 @@ class GrantAccess(APIView):
             user = User.objects.filter(Q(id=user_id) | Q(username=username)).first()
             TeamMembership.objects.filter(user=user,team=team).update(role=type)
             return Response({"message":"success"}, status=200)
-        raise AuthenticationFailed()
+        raise PermissionDenied()
 
 #只有创建者可以撤销权限
 
 class RevokeAccess(APIView):     
-    permission_classes = [IsTeamCreater]
+    permission_classes = [IsAuthenticated]
     def post(self, request):
         if(IsTeamCreater(request)):
             type = request.data.get('type')
@@ -175,10 +187,103 @@ class RevokeAccess(APIView):
             user = User.objects.filter(Q(id=user_id) | Q(username=username)).first()
             TeamMembership.objects.filter(user=user,team=team).update(role=type)
             return Response({"message":"success"}, status=200)
-        raise AuthenticationFailed()
+        raise PermissionDenied()
     
 # class TaskManagerView(viewsets.ModelViewSet):
 
+
+class TaskManage(viewsets.ModelViewSet):
+    permission_classes = [IsAuthenticated]
+    queryset = Task.objects.all()
+    serializer_class = TaskSerializer
+
+    def create(self,request):
+        if(IsTeamMember(request)):
+            # team_id = request.data.pop('team_id',None)
+            serializer = TaskSerializer(data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response({"message":"success"}, status=200)
+            return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
+        raise PermissionDenied()
+    
+    #缺少正在编辑时的项目保护，即有人编辑时应该无法删除
+    @action(detail=False, methods=['delete'])
+    def delete(self,request):
+        if(IsTeamMember(request)):
+            team_id = request.data.get('team_id')
+            task_id = request.data.get('task_id')
+            tasks = Task.objects.filter(team_id=team_id,id=task_id).first()
+            if not tasks:
+                return Response({"message":"task not found"}, status=status.HTTP_400_BAD_REQUEST)
+            #移入垃圾桶
+            dir_path = os.path.join(os.path.abspath('.'),'data','documents',team_id,task_id)
+            recycle_path = os.path.join(os.path.abspath('.'),'recycle',team_id)
+            os.makedirs(recycle_path,exist_ok=True)
+            move_files(dir_path,recycle_path)
+
+            tasks.delete()
+            return Response({"message":"success"}, status=200)
+        raise PermissionDenied()
+    
+
+    
+    
+
+class DocumentManage(viewsets.ModelViewSet):
+    permission_classes = [IsAuthenticated]
+    queryset = Document.objects.all()
+    serializer_class = DocumentSerializer
+
+    def create(self,request):
+        if(IsTeamMember(request)):
+            team_id = request.data.pop('team_id',None)
+            task_id = request.data.get('task_id')
+            dir_path = os.path.join(os.path.abspath('.'),'data','documents',team_id,task_id)
+            document_path = os.path.join(dir_path,''.join([request.data.get('document_name'),'.txt']))
+            request.data['document_path']=document_path
+            request.data['creater_id'] = request.user.id
+            if(os.path.exists(document_path)):
+                    return Response({"message":"document already exists"}, status=status.HTTP_400_BAD_REQUEST)
+            serializer = DocumentSerializer(data=request.data)
+            if serializer.is_valid(): 
+                serializer.save()
+                os.makedirs(dir_path,exist_ok=True)
+                with open(document_path,'w'):
+                    pass
+                return Response({"message":"success"}, status=200)
+            return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
+        raise PermissionDenied()
+    
+    def list(self,request):
+        if(IsTeamMember(request)):
+            team_id = request.data.get('team_id')
+            task_id = request.data.get('task_id')
+            documents = Document.objects.filter(task_id=task_id)
+            return Response(DocumentSerializer(documents,many=True).data,status=200)
+        raise PermissionDenied()
+    
+    #缺少正在编辑时的项目保护，即有人编辑时应该无法删除
+    @action(detail=False, methods=['delete'])
+    def delete(self,request):
+        if(IsTeamMember(request)):
+            team_id = request.data.get('team_id')
+            task_id = request.data.get('task_id')
+            document_id = request.data.get('document_id')
+            documents = Document.objects.filter(task_id=task_id,id=document_id).first()
+            if not documents:
+                return Response({"message":"document not found"}, status=status.HTTP_400_BAD_REQUEST)
+            #移入垃圾桶
+            dir_path = os.path.join(os.path.abspath('.'),'data','documents',team_id,task_id)
+            document_path = os.path.join(dir_path,''.join([request.data.get('document_name'),'.txt']))
+            recycle_path = os.path.join(os.path.abspath('.'),'recycle',team_id)
+            os.makedirs(recycle_path,exist_ok=True)
+            move_file(document_path,recycle_path)
+
+            documents.delete()
+            return Response({"message":"success"}, status=200)
+        raise PermissionDenied()
+
